restructure datastacks cli and generate
