+++ datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/README.md.jinja
@@ -1,0 +1,14 @@
+# Ingest {{ dataset_name }} Workload
+
+__Workload Name:__ ingest_{{ dataset_name }}
+
+__Workload Type:__ Ingest
+
+__Description:__ {{ pipeline_description }}
+
+Created in Stacks Azure Data Platform. Contains resources for:
+
+* Config files
+* Data Factory resources
+* Tests
+* CICD pipelines (Azure DevOps)
