+++ datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/data_factory/adf_pipelines.tf.jinja
@@ -1,0 +1,15 @@
+{% set pipeline_name = 'ingest_' ~ dataset_name -%}
+resource "azurerm_resource_group_template_deployment" "pipeline_{{ pipeline_name }}" {
+  name                = "pipeline_{{ pipeline_name }}"
+  resource_group_name = var.data_factory_resource_group_name
+  deployment_mode     = var.arm_deployment_mode
+  parameters_content = jsonencode({
+    "factoryName" = {
+      value = data.azurerm_data_factory.factory.name
+    }
+  })
+  template_content = file("${path.module}/pipelines/ARM_IngestTemplate.json")
+  depends_on = [
+    azurerm_data_factory_custom_dataset.ds_{{ dataset_name }}
+  ]
+}
