+++ datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/data_factory/provider.tf.jinja
@@ -1,0 +1,8 @@
+terraform {
+  backend "azurerm" {
+  }
+}
+
+provider "azurerm" {
+  features {}
+}
