+++ datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/data_factory/vars.tf.jinja
@@ -1,0 +1,31 @@
+variable "data_factory" {
+  type        = string
+  description = "Azure Data Factory name"
+}
+
+variable "data_factory_resource_group_name" {
+  type        = string
+  description = "Azure Data Factory resource group name"
+}
+
+variable "linked_service_connectionstring" {
+  type        = string
+  description = "Data source connection string, e.g. Integrated Security=False;Encrypt=True;Connection Timeout=30;Data Source=amidostacksdeveuwdesql.database.windows.net;Initial Catalog=sqldbtest;User ID=mssqladmin"
+}
+
+variable "integration_runtime_name" {
+  type        = string
+  description = "Azure Data Factory Integration Runtime name, e.g. adf-managed-vnet-runtime"
+}
+
+variable "key_vault_linked_service_name" {
+  type        = string
+  description = "ADF Key Vault linked service name"
+  default     = "{{ key_vault_linked_service_name }}"
+}
+
+variable "arm_deployment_mode" {
+  type        = string
+  description = "Deployment mode for any ARM resources"
+  default     = "{{ default_arm_deployment_mode }}"
+}
