+++ datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/tests/end_to_end/features/environment.py.jinja
@@ -1,0 +1,14 @@
+{% set pipeline_name = 'ingest_' ~ dataset_name -%}
+from behave import use_fixture
+from behave.model import Scenario
+from behave.runner import Context
+
+from datastacks.behave.fixtures import azure_adls_clean_up
+from datastacks.constants import BRONZE_CONTAINER_NAME
+
+SQL_DB_INGEST_DIRECTORY_NAME = "{{ pipeline_name }}"
+
+
+def before_scenario(context: Context, scenario: Scenario):
+    """Behave before scenario steps."""
+    use_fixture(azure_adls_clean_up, context, BRONZE_CONTAINER_NAME, SQL_DB_INGEST_DIRECTORY_NAME)
