+++ datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/tests/end_to_end/features/steps/azure_data_ingest_steps.py.jinja
@@ -1,0 +1,7 @@
+from datastacks.behave.shared_steps import (  # noqa: F401 flake 8
+    trigger_adf_pipeline,
+    poll_adf_pipeline,
+    pipeline_has_finished_with_state,
+    check_all_files_present_in_adls,
+    check_adf_pipeline_completion_time,
+)
