+++ datastacks/datastacks/generate/templates/processing/processing_template/data_factory/adf_pipelines.tf.jinja
@@ -1,0 +1,11 @@
+resource "azurerm_resource_group_template_deployment" "pipeline_{{ pipeline_name }}" {
+  name                = "pipeline_{{ pipeline_name }}"
+  resource_group_name = var.data_factory_resource_group_name
+  deployment_mode     = var.arm_deployment_mode
+  parameters_content = jsonencode({
+    "factoryName" = {
+      value = data.azurerm_data_factory.factory.name
+    }
+  })
+  template_content = file("${path.module}/pipelines/arm_template.json")
+}
