+++ datastacks/datastacks/generate/templates/processing/processing_template/data_factory/constraints.tf.jinja
@@ -1,0 +1,11 @@
+terraform {
+  required_version = ">= 1.0"
+
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "~> 3.0"
+    }
+
+  }
+}
