+++ datastacks/datastacks/generate/templates/processing/processing_template/data_factory/data.tf.jinja
@@ -1,0 +1,4 @@
+data "azurerm_data_factory" "factory" {
+  name                = var.data_factory
+  resource_group_name = var.data_factory_resource_group_name
+}
