+++ datastacks/datastacks/generate/templates/processing/processing_template/data_factory/provider.tf.jinja
@@ -1,0 +1,8 @@
+terraform {
+  backend "azurerm" {
+  }
+}
+
+provider "azurerm" {
+  features {}
+}
