+++ datastacks/datastacks/generate/templates/processing/processing_template/data_factory/vars.tf.jinja
@@ -1,0 +1,15 @@
+variable "data_factory" {
+  type        = string
+  description = "Azure Data Factory name"
+}
+
+variable "data_factory_resource_group_name" {
+  type        = string
+  description = "Azure Data Factory resource group name"
+}
+
+variable "arm_deployment_mode" {
+  type        = string
+  description = "Deployment mode for any ARM resources"
+  default     = "{{ default_arm_deployment_mode }}"
+}
