+++ datastacks/datastacks/generate/templates/processing/processing_template/tests/unit/conftest.py.jinja
@@ -1,0 +1,15 @@
+from pytest import fixture
+
+from datastacks.pyspark.pyspark_utils import get_spark_session
+
+
+@fixture(scope="session")
+def spark(tmp_path_factory):
+    """Spark session fixture with a temporary directory as a Spark warehouse."""
+    temp_dir = tmp_path_factory.mktemp("spark-warehouse")
+    spark_config = {"spark.sql.warehouse.dir": temp_dir}
+    spark = get_spark_session("{{ pipeline_name }}", spark_config)
+
+    yield spark
+
+    spark.stop()
