+++ datastacks/datastacks/generate/templates/processing/processing_template/tests/unit/test_processing.py.jinja
@@ -1,0 +1,3 @@
+def test_dummy(spark):
+    # Add unit tests for your workload
+    pass
