+++ datastacks/datastacks/generate/templates/processing/processing_template_DQ/README.md.jinja
@@ -1,0 +1,15 @@
+# {{ pipeline_name }} Workload
+
+__Workload Name:__ {{ pipeline_name }}
+
+__Workload Type:__ Processing
+
+__Description:__ {{ pipeline_description }}
+
+Created in Stacks Azure Data Platform. Contains resources for:
+
+* Config files
+* Data Factory resources
+* Spark jobs
+* Tests
+* CICD pipelines (Azure DevOps)
