+++ datastacks/tests/unit/cli/test_cli.py
@@ -3,11 +3,15 @@ import yaml
 from datastacks.cli.datastacks_cli import ingest, processing
 
 
+TEST_CONFIG_DIRECTORY = "datastacks/tests/data/template_config/"
+TEST_CONFIG_INGEST = TEST_CONFIG_DIRECTORY + "test_config_ingest.yml"
+TEST_CONFIG_PROCESS = TEST_CONFIG_DIRECTORY + "test_config_process.yml"
+
+
 def test_cli_ingest():
     runner = CliRunner()
     dq_options = ["--no-data-quality", "--data-quality"]
-    test_config = "datastacks/tests/unit/cli/test_config_ingest.yml"
-    with open(test_config, "r") as file:
+    with open(TEST_CONFIG_INGEST, "r") as file:
         config_dict = yaml.safe_load(file)
 
     with runner.isolated_filesystem():
@@ -22,8 +26,7 @@ def test_cli_ingest():
 def test_cli_processing():
     runner = CliRunner()
     dq_options = ["--no-data-quality", "--data-quality"]
-    test_config = "datastacks/tests/unit/cli/test_config_process.yml"
-    with open(test_config, "r") as file:
+    with open(TEST_CONFIG_PROCESS, "r") as file:
         config_dict = yaml.safe_load(file)
 
     with runner.isolated_filesystem():
