+++ datastacks/tests/unit/generate/test_data_workloads.py
@@ -5,14 +5,19 @@ from unittest.mock import patch
 
 import pytest
 
-from datastacks.cli.config import IngestWorkloadConfigModel
-from datastacks.cli.utils import (
+from datastacks.generate.template_config import IngestWorkloadConfigModel
+from datastacks.generate.data_workloads import (
     validate_yaml_config,
     generate_pipeline,
     generate_target_dir,
     render_template_components,
 )
-from tests.unit.cli.template_structures import EXPECTED_DQ_FILE_LIST, EXPECTED_FILE_LIST
+from tests.unit.generate.conftest import (
+    TEST_CONFIG_INGEST,
+    TEST_CONFIG_INGEST_OVERWRITE,
+    EXPECTED_FILE_LIST,
+    EXPECTED_DQ_FILE_LIST,
+)
 
 
 def test_render_template_components(tmp_path):
@@ -29,7 +34,7 @@ def test_render_template_components(tmp_path):
     }
     config = IngestWorkloadConfigModel(**config_dict)
 
-    template_source_path = "de_templates/ingest/Ingest_SourceType_SourceName/"
+    template_source_path = "datastacks/datastacks/generate/templates/ingest/Ingest_SourceType_SourceName/"
     target_dir = f"{tmp_path}/test_render"
 
     render_template_components(config, template_source_path, target_dir)
@@ -41,30 +46,27 @@ def test_render_template_components(tmp_path):
 @pytest.mark.parametrize(
     "dq,expected_files", [(False, EXPECTED_FILE_LIST), (True, EXPECTED_FILE_LIST + EXPECTED_DQ_FILE_LIST)]
 )
-@patch("datastacks.cli.utils.click.confirm")
-@patch("datastacks.cli.utils.generate_target_dir")
+@patch("datastacks.generate.data_workloads.click.confirm")
+@patch("datastacks.generate.data_workloads.generate_target_dir")
 def test_generate_pipeline(mock_target_dir, mock_confirm, tmp_path, dq, expected_files):
     mock_target_dir.return_value = tmp_path
     mock_confirm.return_value = True
-    config_path = "datastacks/tests/unit/cli/test_config_ingest.yml"
 
-    validated_config = validate_yaml_config(config_path, IngestWorkloadConfigModel)
+    validated_config = validate_yaml_config(TEST_CONFIG_INGEST, IngestWorkloadConfigModel)
     target_dir = generate_pipeline(validated_config, dq)
 
     for file_path in expected_files:
         assert Path(f"{target_dir}/{file_path}").exists()
 
 
-@patch("datastacks.cli.utils.click.confirm")
-@patch("datastacks.cli.utils.generate_target_dir")
+@patch("datastacks.generate.data_workloads.click.confirm")
+@patch("datastacks.generate.data_workloads.generate_target_dir")
 def test_generate_pipeline_new_path(mock_target_dir, mock_confirm, tmp_path):
     mock_target_dir.return_value = tmp_path
     mock_confirm.return_value = False
     rmtree(tmp_path)
 
-    config_path = "datastacks/tests/unit/cli/test_config_ingest.yml"
-
-    validated_config = validate_yaml_config(config_path, IngestWorkloadConfigModel)
+    validated_config = validate_yaml_config(TEST_CONFIG_INGEST, IngestWorkloadConfigModel)
     target_dir = generate_pipeline(validated_config, False)
 
     for file_path in EXPECTED_FILE_LIST:
@@ -74,15 +76,13 @@ def test_generate_pipeline_new_path(mock_target_dir, mock_confirm, tmp_path):
 @pytest.mark.parametrize(
     "overwrite_confirm,expected_desc", [(False, "Pipeline for testing"), (True, "Pipeline for testing overwritten")]
 )
-@patch("datastacks.cli.utils.click.confirm")
-@patch("datastacks.cli.utils.generate_target_dir")
+@patch("datastacks.generate.data_workloads.click.confirm")
+@patch("datastacks.generate.data_workloads.generate_target_dir")
 def test_generate_pipeline_overwrite(mock_target_dir, mock_confirm, tmp_path, overwrite_confirm, expected_desc):
     mock_target_dir.return_value = tmp_path
     mock_confirm.return_value = True
 
-    config_path = "datastacks/tests/unit/cli/test_config_ingest.yml"
-
-    validated_config = validate_yaml_config(config_path, IngestWorkloadConfigModel)
+    validated_config = validate_yaml_config(TEST_CONFIG_INGEST, IngestWorkloadConfigModel)
     target_dir = generate_pipeline(validated_config, False)
 
     for file_path in EXPECTED_FILE_LIST:
@@ -92,10 +92,9 @@ def test_generate_pipeline_overwrite(mock_target_dir, mock_confirm, tmp_path, ov
         arm_template_dict = json.load(file)
     assert arm_template_dict["resources"][0]["properties"]["description"] == "Pipeline for testing"
 
-    config_path = "datastacks/tests/unit/cli/test_config_ingest_overwrite.yml"
     mock_confirm.return_value = overwrite_confirm
 
-    validated_config = validate_yaml_config(config_path, IngestWorkloadConfigModel)
+    validated_config = validate_yaml_config(TEST_CONFIG_INGEST_OVERWRITE, IngestWorkloadConfigModel)
     target_dir = generate_pipeline(validated_config, False)
 
     with open(f"{target_dir}/data_factory/pipelines/ARM_IngestTemplate.json") as file:
@@ -103,15 +102,13 @@ def test_generate_pipeline_overwrite(mock_target_dir, mock_confirm, tmp_path, ov
     assert arm_template_dict["resources"][0]["properties"]["description"] == expected_desc
 
 
-@patch("datastacks.cli.utils.click.confirm")
-@patch("datastacks.cli.utils.generate_target_dir")
+@patch("datastacks.generate.data_workloads.click.confirm")
+@patch("datastacks.generate.data_workloads.generate_target_dir")
 def test_enum_templating(mock_target_dir, mock_confirm, tmp_path):
     mock_target_dir.return_value = tmp_path
     mock_confirm.return_value = True
 
-    config_path = "datastacks/tests/unit/cli/test_config_ingest.yml"
-
-    validated_config = validate_yaml_config(config_path, IngestWorkloadConfigModel)
+    validated_config = validate_yaml_config(TEST_CONFIG_INGEST, IngestWorkloadConfigModel)
     target_dir = generate_pipeline(validated_config, False)
 
     tested_file_path = f"{target_dir}/data_factory/pipelines/ARM_IngestTemplate.json"
