+++ de_workloads/generate_examples/README.md
@@ -1,0 +1,5 @@
+# Generate workload examples
+
+The files in this directory are example config files to be used for generating data workloads using Datastacks.
+
+See the [documentation](http://stacks.ensono.com/docs/workloads/azure/data/data_engineering/datastacks) for further details.
